LPS-51081 No need for an internal package
